pyproject.toml
[build-system]
requires = ["setuptools>=80.1.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "astra"
version = "0.1.0"
description = "凌云曦(Astra): 基于Telegram的多AI智能体"
authors = [{name="Eoysky", email="eoysky@eoysky.com"}]
requires-python = ">=3.13"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "python-telegram-bot==22.0",
    "APScheduler==3.11.0",
    "aiohttp==3.11.18",
    "python-dotenv==1.1.0",
    "httpx==0.28.1",
    "requests==2.32.3",
    "toml==0.10.2",
    "yarl==1.20.0",
    "aiolimiter==1.2.1",
    "cachetools==5.5.2",
    "psycopg2-binary==2.9.10",
    "attrs==25.3.0",
    "pytest==8.3.5",
    "pytest-cov==6.1.1",
    "coverage==7.8.0",
    "multidict==6.4.3",
    "tzlocal==5.3.1",
    "cryptography==44.0.3",
    "cffi==1.17.1",
    "pycparser==2.22",
    "anyio==4.9.0",
    "sniffio==1.3.1",
    "frozenlist==1.6.0",
    "aiohappyeyeballs==2.6.1",
    "h11==0.16.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "hyperframe==6.1.0",
    "iniconfig==2.1.0",
    "packaging==25.0",
    "pluggy==1.5.0",
    "propcache==0.3.1",
    "charset-normalizer==3.4.2",
    "certifi==2025.4.26",
    "socksio==1.0.0",
    "build==1.2.2.post1",
    "pyproject_hooks==1.2.0",
    "setuptools==80.1.0",
    "wheel==0.45.1",
    "tornado==6.4.2",
    "idna==3.10",
    "asyncio==3.4.3"
]
